requirements.txt
# Core Data Science
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.10.1

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.15.0

# ML Ops
mlflow==2.5.0
airflow==2.6.3

# API
fastapi==0.100.0
uvicorn==0.23.1
pydantic==2.0.3

# Database
sqlalchemy==2.0.19
sqlite3

# Utils
pyyaml==6.0
python-dotenv==1.0.0
click==8.1.6

# Development
pytest==7.4.0
black==23.7.0
flake8==6.0.0
pre-commit==3.3.3

# Documentation
mkdocs==1.4.3
mkdocs-material==9.1.19
